Authority Console Phase 5: runs tab + final cleanup (initiative complete)
Absorb the last standalone panel into the console and remove all dead code.

- New RunsTab re-mounts the existing EnrichmentRunner / EnrichmentJobList
  (useOptimisticRows) unchanged inside the console — those stay in
  components/admin/enrichment/ because CorpusEnrichmentCard also uses them; only
  AdminEnrichment's page wrapper is absorbed.
- DELETED AdminEnrichment.tsx + its CT + wrapper + /admin/enrichment route +
  export, and the transitional PlaceholderTab (all five tabs are now real).

This completes the unified Authority Console: the three former standalone admin
panels (/admin/authorities, /admin/authority-mappings, /admin/enrichment) are all
deleted, collapsed into one front door at /admin/authority with tabs for
Authorities · Aliases & Relationships · Discovery Queue · Scrapers · Runs.
Zero duplicate components, zero dead code.

tsc + prettier clean; 8 Playwright CT green (all five tabs covered).
