+++ frontend/src/components/admin/index.ts
@@ -3,6 +3,5 @@ export { GlobalAgentManagement } from "./GlobalAgentManagement";
 export { SystemSettings } from "./SystemSettings";
 export { WorkerAccountManagement } from "./WorkerAccountManagement";
 export { IngestionMonitor } from "./IngestionMonitor";
-export { AdminEnrichment } from "./AdminEnrichment";
 export { AuthorityConsole } from "./authority/AuthorityConsole";
 export { CorpusCategoryManagement } from "./corpus_categories/CorpusCategoryManagement";
